+++ jitsi_meet_wrapper/example/lib/main.dart
@@ -161,11 +161,49 @@ class _MeetingState extends State<Meeting> {
       options: options,
       listener: JitsiMeetingListener(
         onOpened: () => debugPrint("onOpened"),
-        onConferenceWillJoin: (url) => debugPrint("onConferenceWillJoin: url: $url"),
-        onConferenceJoined: (url) => debugPrint("onConferenceJoined: url: $url"),
+        onConferenceWillJoin: (url) {
+          debugPrint("onConferenceWillJoin: url: $url");
+        },
+        onConferenceJoined: (url) {
+          debugPrint("onConferenceJoined: url: $url");
+        },
         onConferenceTerminated: (url, error) {
           debugPrint("onConferenceTerminated: url: $url, error: $error");
         },
+        onAudioMutedChanged: (isMuted) {
+          debugPrint("onAudioMutedChanged: isMuted: $isMuted");
+        },
+        onVideoMutedChanged: (isMuted) {
+          debugPrint("onVideoMutedChanged: isMuted: $isMuted");
+        },
+        onScreenShareToggled: (participantId, isSharing) {
+          debugPrint(
+            "onScreenShareToggled: participantId: $participantId, "
+            "isSharing: $isSharing",
+          );
+        },
+        onParticipantJoined: (email, name, role, participantId) {
+          debugPrint(
+            "onParticipantJoined: email: $email, name: $name, role: $role, "
+            "participantId: $participantId",
+          );
+        },
+        onParticipantLeft: (participantId) {
+          debugPrint("onParticipantLeft: participantId: $participantId");
+        },
+        onParticipantsInfoRetrieved: (participantsInfo, requestId) {
+          debugPrint(
+            "onParticipantsInfoRetrieved: participantsInfo: $participantsInfo, "
+            "requestId: $requestId",
+          );
+        },
+        onChatMessageReceived: (senderId, message, isPrivate) {
+          debugPrint(
+            "onChatMessageReceived: senderId: $senderId, message: $message, "
+            "isPrivate: $isPrivate",
+          );
+        },
+        onChatToggled: (isOpen) => debugPrint("onChatToggled: isOpen: $isOpen"),
         onClosed: () => debugPrint("onClosed"),
       ),
     );