+++ jitsi_meet_wrapper/CHANGELOG.md
@@ -1,8 +1,21 @@
+## 0.0.4
+
+- Bump SDK version
+  to [5.0.2](https://github.com/jitsi/jitsi-meet-release-notes/blob/master/CHANGELOG-MOBILE-SDKS.md#502-2022-03-29) [#19](https://github.com/saibotma/jitsi_meet_wrapper/pull/19)
+- Add more conference events: `onAudioMutedChanged`, `onVideoMutedChanged`, `onScreenShareToggled`
+  , `onParticipantJoined`, `onParticipantLeft`,  `onParticipantsInfoRetrieved`, `onChatMessageReceived`
+  , `onChatToggled` [#18](https://github.com/saibotma/jitsi_meet_wrapper/pull/18)
+
 ## 0.0.3
-- Bump SDK version to [5.0.0](https://github.com/jitsi/jitsi-meet-release-notes/blob/master/CHANGELOG-MOBILE-SDKS.md#500-2022-03-02)
+
+- Bump SDK version
+  to [5.0.0](https://github.com/jitsi/jitsi-meet-release-notes/blob/master/CHANGELOG-MOBILE-SDKS.md#500-2022-03-02)
 
 ## 0.0.2
-- Bump SDK version to [4.1.0](https://github.com/jitsi/jitsi-meet-release-notes/blob/master/CHANGELOG-MOBILE-SDKS.md#410-2021-12-14)
+
+- Bump SDK version
+  to [4.1.0](https://github.com/jitsi/jitsi-meet-release-notes/blob/master/CHANGELOG-MOBILE-SDKS.md#410-2021-12-14)
 
 ## 0.0.1
+
 - Initial version