+++ jitsi_meet_wrapper_platform_interface/CHANGELOG.md
@@ -1,2 +1,9 @@
+## 0.0.2
+
+- Add more conference events: `onAudioMutedChanged`, `onVideoMutedChanged`, `onScreenShareToggled`
+  , `onParticipantJoined`, `onParticipantLeft`,  `onParticipantsInfoRetrieved`, `onChatMessageReceived`
+  , `onChatToggled` [#18](https://github.com/saibotma/jitsi_meet_wrapper/pull/18)
+
 ## 0.0.1
+
 - Initial version